bot/requirements.txt
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-PyMongo==2.3.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.16.0
lxml==4.5.1
MarkupSafe==1.1.1
Pillow==7.1.2
pymongo==3.10.1
python-dotenv==0.13.0
pyzbar==0.1.8
requests==2.23.0
soupsieve==2.0.1
urllib3==1.25.9
uWSGI==2.0.19.1
Werkzeug==1.0.1
wikipedia==1.4.0
